Further updated the Renderer.
Instead of passing all quads to the gpu every frame, the renderer now just replaces the data that has changed. It increased performance by about 25% in the test environment, but there may be performance issues down the line when the almost everything changes position every frame. This is still not completely optimized.