Added more default shaders for use based on system specs.
Was running into issues with the default fragment shader being hardcoded to 80 texture slots which was no problem with a modern dedicated gpu, but was too many for integrated graphics. Changed that number to 16 which is the minimum required by OpenGL 4.6. Also added another fragment shader for systems that support at least 32 texture slots. On initialization, the renderer checks the number of texture slots of the system and selects the appropriate fragment shader. The goal is to add more default shaders for systems with different texture slots.