Setup minimum pieces to make a scene tree.
Primarily just created a way to connect entities together and attach them to a scene. This is not completely finished as components among many other things still need to be added and it also needs to be tested. The only testing that was done was quite crude and far from extensive. This is a key aspect of Sprocket so it should be heavily tested. It should also be looked over to ensure pointers, references, and const value are being utilized as intended.