Implemented a semi-functional component system to entities.
Currently it does not consistently pass the tests written for it, so it needs to be examined further to get it properly written. In the test component, the stored value is not inconsistent during testing. Further testing needs to be done to find the issue. Once the issue is found and solved, the ECS will be ready to have more component types introduced.