Solved the issue with the components.
Was using the Component copy constructor instead of the one for the specific component type which meant that the internal component specific values were not being copied. Solved the problem using a switch statement on the component type and using a different copy constructor for each component type. There may be a better way of doing this, so that should be explored in the future.