Created render events.
This allows calls to the renderer to be done through the event system instead of directly. This allows the systems be stay loosely connected instead of having direct calls. This system is a bit difficult to understand exactly how it works so good documentation for it is a priority. Now more than ever there needs to be some kind of distinction as far as what event types subscribers are subscribing to because it does not make sense to propogate render events to things like the window and input system.