Added a QuadRenderer Component.
This is not a full featured, finished component, but is functional enough that rendering can now be done by simply adding a QuadRenderer to a component in an active scene. Before any further progress is made on this, there needs to be refactoring on some systems because the interaction is a bit messy.