Seperated the component functionality into a seperate system.
This lets components be purely a data storage container that is utilized by various systems. This approach seems to allow better decoupling that the previous iteration. There still needs to be a solution for transfering the data to the various systems, as right now it has to be done manually in the user applicaion. So far this implementation is much easier to understand and work with, which will hopefully allow it to scale better.