Got quad rendering to be almost full encapsulated by ECS.
Now quads can be rendered simply by creating entities and attaching quadrenderercomponents to them. The current implementation to check for modifications to component data is quite expensive and in a stress test, performance was less than ideal. It is possible that the problem is not from this ECS system, but it is the most likely culprit. Profiling needs to be done to be sure. There also needs to be a limit of 1 component per type per entity to prevent an entity from containing tons of overlapping quads. If stacking quads in needed, one can instead add a child component with a zeroed out local transform.