Limited components to 1 per type per entity.
Also added the ablility to change a textured quad into a plain color quad, since there inverse was already possible. Also, upon further inspection of the SceneManager's check for modified components, it seems that the current implementation may be okay. Perhaps being able to mark an entity as static would be a good idea in the future, that way there is a way to differentiate the entities that will not have modified data without having to check all the data.